feat: post controller

+++ NestFitMvc/src/main/java/org/aimable/nestfitmvc/DAO/PostDAO.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface PostDAO {
     void createPost(Post post) throws SQLException;
     List<Post> getAllPosts();
-    Post getPostById(int id) throws SQLException;
+    Post getPostById(Long id) throws SQLException;
     void updatePost(Post post) throws SQLException;
-    void deletePost(int id) throws SQLException;
+    void deletePost(Long id) throws SQLException;
 }
